pyproject.toml
[build-system]
requires = ["flit_core>=3.2", "flit_scm", "wheel"]
build-backend = "flit_scm:buildapi"

[project]
name = "django-oscar-stripe-sca"
version = "0.8.2-3"
description = "Stripe Checkout payment module for django-oscar"
readme = "README.rst"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    {name = "Steve Bradshaw", email = "steve@pcfive.co.uk"},
    {name = "Mathieu Richardoz", email = "richardoz@nngroup.com"},
]
keywords = [
    "django",
    "oscar",
    "payment",
    "stripe",
]
classifiers = [ 
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3 :: Only"
]
dependencies = [ 
    "django>=3.2",
    "django-oscar>=2",
    "stripe>=2.27.0",
]

[project.urls]
Homepage = "https://github.com/st8st8/django-oscar-stripe-sca"

[tool.flit.module]
name = "oscar_stripe_sca"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
exclude = [ "sandbox*",]
namespaces = false
